Phanthom X/requirements.txt
rich
pyyaml

